Deleted redundant files and flattened hierarchy
