+++ webapp/src/components/admin/specialization-index/SpecializationIndexList.tsx
@@ -2,13 +2,13 @@ import { Button, makeStyles, shorthands, tokens } from '@fluentui/react-componen
 import { FC } from 'react';
 import { DndProvider } from 'react-dnd';
 import { HTML5Backend } from 'react-dnd-html5-backend';
-import { useSpecialization } from '../../../../libs/hooks';
-import { useAppDispatch, useAppSelector } from '../../../../redux/app/hooks';
-import { RootState } from '../../../../redux/app/store';
-import { setSelectedKey } from '../../../../redux/features/admin/adminSlice';
-import { Breakpoints } from '../../../../styles';
-import { Add20 } from '../../../shared/BundledIcons';
-import { AdminItemListSection } from '../../shared/list/AdminListItemSection';
+import { useSpecializationIndex } from '../../../libs/hooks/useSpecializationIndex';
+import { useAppDispatch, useAppSelector } from '../../../redux/app/hooks';
+import { RootState } from '../../../redux/app/store';
+import { setSelectedIndexKey } from '../../../redux/features/admin/adminSlice';
+import { Breakpoints } from '../../../styles';
+import { Add20 } from '../../shared/BundledIcons';
+import { AdminItemListSection } from '../shared/list/AdminListItemSection';
 
 const useClasses = makeStyles({
     root: {
@@ -16,8 +16,8 @@ const useClasses = makeStyles({
         flexShrink: 0,
         width: '320px',
         backgroundColor: tokens.colorNeutralBackground4,
-        flexDirection: 'column',
         boxShadow: 'rgba(0, 0, 0, 0.25) 0px 0.2rem 0.4rem -0.075rem',
+        flexDirection: 'column',
         ...shorthands.overflow('hidden'),
         ...Breakpoints.small({
             width: '64px',
@@ -51,13 +51,13 @@ const useClasses = makeStyles({
     },
 });
 
-export const SpecializationList: FC = () => {
+export const SpecializationIndexList: FC = () => {
     const classes = useClasses();
     const dispatch = useAppDispatch();
-    const specialization = useSpecialization();
-    const { specializations, selectedId } = useAppSelector((state: RootState) => state.admin);
+    const indexes = useSpecializationIndex();
+    const { specializationIndexes, selectedIndexId } = useAppSelector((state: RootState) => state.admin);
     const onAddSpecializationClick = () => {
-        dispatch(setSelectedKey(''));
+        dispatch(setSelectedIndexKey(''));
     };
 
     return (
@@ -73,20 +73,20 @@ export const SpecializationList: FC = () => {
                                 onAddSpecializationClick();
                             }}
                         >
-                            New Specialization
+                            New Index
                         </Button>
                     </div>
-                    <div aria-label={'specialization list'} className={classes.list}>
+                    <div aria-label={'specialization index list'} className={classes.list}>
                         <AdminItemListSection
-                            items={specializations}
-                            onItemCollectionReorder={(specs) => {
-                                void specialization.setSpecializationsOrder(specs);
+                            header="All"
+                            items={specializationIndexes}
+                            onItemCollectionReorder={(items) => {
+                                void indexes.setSpecializationIndexOrder(items);
                             }}
-                            onItemSelected={(id) => dispatch(setSelectedKey(id))}
-                            selectedId={selectedId}
-                            onItemToggled={(id, toggle) => {
-                                return specialization.toggleSpecialization(id, toggle);
+                            onItemSelected={(id: string) => {
+                                dispatch(setSelectedIndexKey(id));
                             }}
+                            selectedId={selectedIndexId}
                         />
                     </div>
                 </div>
