--- webapp/src/components/admin/specialization-index/SpecializationIndexWindow.tsx
@@ -1,24 +1,0 @@
-import { makeStyles, tokens } from '@fluentui/react-components';
-
-import React from 'react';
-import { SpecializationIndexManager } from './SpecializationIndexManager';
-
-const useClasses = makeStyles({
-    root: {
-        display: 'flex',
-        flexDirection: 'column',
-        width: '100%',
-        backgroundColor: tokens.colorNeutralBackground3,
-        boxShadow: 'rgb(0 0 0 / 25%) 0 0.2rem 0.4rem -0.075rem',
-    },
-});
-
-export const SpecializationIndexWindow: React.FC = () => {
-    const classes = useClasses();
-
-    return (
-        <div className={classes.root}>
-            <SpecializationIndexManager />
-        </div>
-    );
-};
