+++ webapp/src/components/views/ChatView.tsx
@@ -3,10 +3,10 @@ import { FC } from 'react';
 import { useAppSelector } from '../../redux/app/hooks';
 import { RootState } from '../../redux/app/store';
 import { AdminWindow } from '../admin/shared/AdminWindow';
+import { SpecializationIndexList } from '../admin/specialization-index/SpecializationIndexList';
 import { SpecializationIndexManager } from '../admin/specialization-index/SpecializationIndexManager';
-import { SpecializationIndexList } from '../admin/specialization-index/index-list/SpecializationIndexList';
+import { SpecializationList } from '../admin/specialization/SpecializationList';
 import { SpecializationManager } from '../admin/specialization/SpecializationManager';
-import { SpecializationList } from '../admin/specialization/specialization-list/SpecializationList';
 import { ChatWindow } from '../chat/ChatWindow';
 import { ChatType } from '../chat/chat-list/ChatType';
 import { SearchWindow } from '../search/SearchWindow';
